rename to extractor
